Refactor : read와 commad 서비스 분리
