+++ core/src/main/java/com/garden/back/garden/service/GardenReadService.java
@@ -68,18 +68,4 @@ public class GardenReadService {
         return RecentGardenResults.to(gardenHistoryManager.getRecentViewGarden(memberId));
     }
 
-    @Transactional
-    public void deleteGarden(GardenDeleteParam param) {
-        Garden gardenToDelete = gardenRepository.getById(param.gardenId());
-
-        if (!Objects.equals(gardenToDelete.getWriterId(), param.memberId())) {
-            throw new IllegalArgumentException("텃밭 게시글 작성자만 텃밭을 삭제할 수 있습니다.");
-        }
-
-        gardenImageRepository.deleteByGardenId(param.gardenId());
-        gardenRepository.deleteById(param.gardenId());
-
-    }
-
-
 }
